new version of composite
