+++ Anvov/Composite/Api/Unit.php
@@ -1,0 +1,19 @@
+<?php
+/**
+ * Composit
+ *
+ * @category school5
+ * @package Adapter
+ * @author Andriy Vovk <andriy.vovk@smile-ukraine.com>
+ * @copyright 2019 Smile
+ */
+interface Unit
+{
+    /**
+     * general method for all elements
+     *
+     * @return mixed
+     *
+     */
+  public function bombardStrenghth();
+}
